chore(wmg): refactor wmg2 main
