+++ frontend/src/views/WheresMyGeneV2/components/Main/constants.ts
@@ -1,0 +1,2 @@
+export const GENGE_EXPRESSION_PAGE_TITLE =
+  "Gene Expression - CZ CELLxGENE Discover";
